"Refactor Footer and Header components: remove SearchBar and cards-related logic from Footer, add SearchBar and cards-related logic to Header, and update styles and conditionals."
